refactor: eliminate the parser
